Move fit tests to datasets/tests
